pyproject.toml
[tool.poetry]
name = "api_rest"
version = "0.1.0"
description = ""
authors = ["InspiredCoconut <kricaldezfernandez@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
poethepoet = "^0.31.1"
poetry-plugin-export = "^1.8.0"
drf-yasg = "^1.21.8"

[tool.poe.tasks]
runserver = "python manage.py runserver"
check = "python manage.py check"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
createsuperuser = "python manage.py createsuperuser"
export = "poetry export -f requirements.txt --output requirements.txt --without-hashes"
import = "poetry add $(cat requirements.txt | tr '\n' ' ')"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
